+++ document.lib.rest/Api/FolderApiService.cs
@@ -1,10 +1,10 @@
 ﻿using System.Linq.Expressions;
-using document.lib.rest.Parameters;
+using document.lib.rest.Api.Contract;
+using document.lib.rest.Interfaces;
 using document.lib.shared.Interfaces;
 using document.lib.shared.Models.Models;
-using Microsoft.AspNetCore.Components.Forms;
 
-namespace document.lib.rest.ApiServices;
+namespace document.lib.rest.Api;
 
 internal class FolderApiService(IFolderService folderService)
 {
@@ -16,8 +16,8 @@ internal class FolderApiService(IFolderService folderService)
     public async Task<IResult> GetFolderModel(FolderGetQueryParameters folderGetQueryParameters, HttpContext http)
     {
         try
-        {
-            if (CheckForNull(folderGetQueryParameters, 
+        {   
+            if (PropertyValidator.ValidateHasValue(folderGetQueryParameters, 
                     x => x.Page, 
                     x => x.PageSize))
             {
@@ -25,16 +25,13 @@ internal class FolderApiService(IFolderService folderService)
                 http.Response.Headers.Append("total-results", count.ToString());
                 return Results.Ok(folders);
             }
-
-            
-            
-
+                
             FolderModel? folder = null;
-            if (!string.IsNullOrWhiteSpace(folderGetQueryParameters.Id?.ToString()))
+            if (PropertyValidator.ValidateHasValue(folderGetQueryParameters, x => x.Id))
                 folder = await folderService.GetFolderByIdAsync(folderGetQueryParameters.Id.ToString()!);
 
-            if (!string.IsNullOrWhiteSpace(folderGetQueryParameters.Name))
-                folder = await folderService.GetFolderByNameAsync(folderGetQueryParameters.Name);
+            if (PropertyValidator.ValidateHasValue(folderGetQueryParameters, x => x.Name))
+                folder = await folderService.GetFolderByNameAsync(folderGetQueryParameters.Name!);
 
             if (folder == null) 
                 return Results.NotFound(folderGetQueryParameters);
@@ -46,21 +43,6 @@ internal class FolderApiService(IFolderService folderService)
             // TODO: logging
             return Results.StatusCode(500);
         }
-    }
-
-    private bool CheckForNull(FolderGetQueryParameters parameters, params Expression<Func<FolderGetQueryParameters, int?>>[] expressions)
-    {
-        List<bool> results = [];
-        foreach (var expression in expressions)
-        {
-            if (expression.Body is MemberExpression body)
-            {
-                var propValue = expression.Compile()(parameters);
-                results.Add(propValue.HasValue);
-            }
-        }
-        
-        return results.All(x => x);
     }
 
     public async Task<IResult> CreateFolder(FolderPutParameters folderPutParameters)
